fix: 브랜드, 영화관 동시성 처리 추가
