+++ src/main/java/com/cinefinder/global/config/DataInitConfig.java
@@ -6,11 +6,17 @@ import com.cinefinder.theater.data.repository.BrandRepository;
 import com.cinefinder.theater.service.TheaterService;
 import jakarta.annotation.PostConstruct;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.hibernate.exception.ConstraintViolationException;
 import org.springframework.beans.factory.annotation.Value;
+import org.springframework.dao.CannotAcquireLockException;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Component;
 
+import java.sql.SQLIntegrityConstraintViolationException;
 import java.util.List;
 
+@Slf4j
 @Component
 @RequiredArgsConstructor
 public class DataInitConfig {
@@ -36,7 +42,12 @@ public class DataInitConfig {
                     Brand.builder().name(lotteBrandName).build(),
                     Brand.builder().name(megaBrandName).build()
             );
-            brandRepository.saveAll(initialBrands);
+            try {
+                brandRepository.saveAll(initialBrands);
+                log.info("✅[브랜드 초기화] 초기 브랜드를 저장했습니다.");
+            } catch (ConstraintViolationException | DataIntegrityViolationException e) {
+                log.warn("⚠️[브랜드 초기화] 다른 서버에서 브랜드를 이미 처리하고 있어서 초기화를 스킵합니다.");
+            }
         }
 
         theaterService.getTheaterInfosAfterSync();
