+++ src/main/java/com/cinefinder/theater/data/Brand.java
@@ -11,13 +11,13 @@ import lombok.NoArgsConstructor;
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-@Table(name = "BRAND")
+@Table(name = "BRAND", uniqueConstraints = {@UniqueConstraint(columnNames = "name")})
 public class Brand {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    @Column(nullable = false)
+    @Column(nullable = false, unique = true)
     private String name;
 }
